+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added method `Coord2.distance()`
+  - Calculates euclidean or taxicab distance between two coordinates depending on the `mode` parameter
+
 ### Changed
 
 ### Deprecated